Extract recipe_info() and practice_recipe_description() to helpers
Move the recipe info text builders out of crafting_gui.cpp into
crafting_gui_helpers so both the curses UI and the future ImGui
implementation can share them.

- recipe_info() builds the recipe info panel text
- practice_recipe_description() builds practice recipe text
- craft_success_chance_string() and cata_fail_chance_string() move as
  file-local helpers in the .cpp
- flag_BLIND_EASY / flag_BLIND_HARD inlined as string literals
- practice_recipe_description() now guards against null practice_data
- test_tallow gets book_learn for "Written in:" test coverage
- New test cases covering recipe info output, practice descriptions,
  NPC helper "Known by:" branch, and book source "Written in:" branch
